requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
configparser==3.5.0
cryptography==2.3.1
Django==2.1.2
djangorestframework==3.8.2
idna==2.7
pycparser==2.19
PyMySQL==0.9.2
pytz==2018.5
six==1.11.0
